Move ElementHandle tests into a single file
There's no reason each method should be in a separate test file, and the
reasoning behind which should or shouldn't isn't clear.
